+++ src/constants/localStorageKeys.js
@@ -1,5 +1,4 @@
 /**
- * @constant {string} PROJECTS_KEY
  * This constant stores the name of the key
  * that is used to write data about the projects in localStorage
  */
