+++ src/pages/SignUp.vue
@@ -19,7 +19,7 @@
 <script>
 import Form from '../components/Form';
 import { SIGN_UP_REQUEST } from '../store/actions/auth';
-import { offlineErrorMessage } from '../mixins/offline-error-message';
+import { offlineErrorMessage } from '../mixins/offlineErrorMessage';
 
 export default {
   name: 'SignUp',
